RD file selection now consistent and partial season releases should be handled
Realdebrid keeps multiple versions of cached torrents, that each contain a different selection of the torrents files. 

Its a stupid system. To download a torrent directly, you need to select the same files as one of the cached versions.

Ive added a ranking system for those cached versions that ranks each version by the amount of files we want (e.g. all released episodes of a season) and by the amount of files we dont want (e.g. sample files, nfo files, etc.).

It then tries to download the best version. The problem is, that sometimes realdebrid makes you download a .rar archive if you select the wrong cached version. Its a really stupid system. Ive added a check for that too.

You will now download the cached version of the torrent that has the most wanted, least unwanted and least files overall that isnt packed into an archive. Jesus.