added basic prowlarr integration
added support for prowlarr as an indexer manager. Sadly, the way prowlarr works does not allow for asynchronous download requests - which basically means that unless a prowlarr indexer provides a releases magnet-link directly, only a few releases from that indexer will be shown until the timeout of 2 seconds is reached. I will try to make this work better, but Im not sure its possible.
Rarbg works fine, but 1337x for example will only show 1 or 2 releases before the request to prowlarr times out.